test_store/requirements.txt
apipkg==1.4
appdirs==1.4.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
configparser==3.5.0
coverage==4.3.4
decorator==4.0.11
Django==1.10.5
django-appconf==1.0.2
django-compressor==2.1.1
django-debug-toolbar==1.6
django-environ==0.4.1
django-extensions==1.7.6
django-extra-views==0.6.4
django-haystack==2.5.1
django-oscar==1.4
django-phonenumber-field==1.1.0
django-redis==4.7.0
django-tables2==1.3.0
django-treebeard==4.1.0
django-webtest==1.8.0
django-widget-tweaks==1.4.1
enum34==1.1.6
execnet==1.4.1
factory-boy==2.8.1
Faker==0.7.7
flake8==3.3.0
flake8-blind-except==0.1.1
flake8-debugger==1.4.0
funcsigs==1.0.2
ipaddress==1.0.18
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
isort==4.2.5
mccabe==0.6.1
mock==2.0.0
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pbr==1.10.0
pexpect==4.2.1
phonenumbers==8.2.0
phonenumberslite==8.2.0
pickleshare==0.7.4
Pillow==4.0.0
pluggy==0.3.1
prompt-toolkit==1.0.13
ptyprocess==0.5.1
purl==1.3
py==1.4.32
pycodestyle==2.3.1
pycountry==17.1.8
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.1.10
pyprof2calltree==1.4.0
pysolr==3.6.0
pysqlite==2.6.3
pytest==3.0.6
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-warnings==0.2.0
pytest-xdist==1.15.0
python-dateutil==2.6.0
pytz==2016.10
rcssmin==1.0.6
redis==2.10.5
requests==2.13.0
rjsmin==1.0.12
scandir==1.4
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.4a1
sqlparse==0.2.2
tox==2.3.1
traitlets==4.3.1
transifex-client==0.12.4
Unidecode==0.4.20
urllib3==1.20
uWSGI==2.0.14
virtualenv==15.1.0
waitress==1.0.2
wcwidth==0.1.7
WebOb==1.7.1
WebTest==2.0.25
Werkzeug==0.11.15
whitenoise==3.3.0
Whoosh==2.7.4
